+++ src/squanmate/scramblers/shape_scrambler.cljs
@@ -32,7 +32,7 @@
     "Select available shapes for scrambles by filtering:"]
    [layer-selector/layer-selector state]])
 
-(defn settings [state]
+(defn settings-component [state]
   [common/accordion {:default-active-key 1}
    [common/panel {:header (reagent/as-element [:span [common/glyphicon {:glyph :cog}]
                                                " Cases"])
@@ -41,24 +41,18 @@
    [common/panel {:header (reagent/as-element [:span [common/glyphicon {:glyph :wrench}]
                                                " Scramble options"])
                   :event-key 2}
-    [middle-layer-controls/controls (reagent/cursor state [:middle-layer-settings])]]
-   [common/panel {:header (reagent/as-element [:span [common/glyphicon {:glyph :tint}]
-                                               " Colors"])
-                  :event-key 3}
-    [color-chooser/color-chooser (reagent/cursor state [:draw-settings])]]])
+    [middle-layer-controls/controls (reagent/cursor state [:middle-layer-settings])]]])
 
 ;; let this module own its state schema by having it defined inside this file
 (defn new-state []
   ;; Allowed shapes are stored as sets of their layers. This makes adding and
   ;; removing them very easy in code.
-  (let [state (reagent/atom
-               {:puzzle nil
-                :chosen-shapes nil
-                :selected-shapes #{(set ["square" "square"])}
-                :scramble-algorithm nil
-                :draw-settings (deref (color-chooser/default-color-chooser-state))
-                :middle-layer-settings (deref (middle-layer-controls/default-state))})]
-    state))
+  (reagent/atom
+   {:puzzle nil
+    :chosen-shapes nil
+    :selected-shapes #{(set ["square" "square"])}
+    :scramble-algorithm nil
+    :middle-layer-settings (deref (middle-layer-controls/default-state))}))
 
 (defn- repeat-case-button [state]
   [common/split-button {:on-click #(a/set-new-repeat-scramble state)
@@ -95,14 +89,17 @@
     [newmonochrome/monochrome-puzzle p draw-settings]
     [common/alert "Select some cases and click 'New scramble' to get started."]))
 
-(defn scramble-component [state]
-  (let [draw-settings (assoc (:draw-settings @state)
-                             :size 180)]
-    [:div
-     [:div.bottom17
-      [action-buttons state]]
-     [:div.center
-      [puzzle-preview state draw-settings]]
-     [:div.center
-      [scramble-preview (:scramble-algorithm @state)]]
-     [settings state]]))
+(defn scramble-component
+  ([state]
+   [scramble-component state newmonochrome/default-settings])
+  ([state draw-settings-map]
+   (let [draw-settings (assoc draw-settings-map
+                              :size 180)]
+     [:div
+      [:div.bottom17
+       [action-buttons state]]
+      [:div.center
+       [puzzle-preview state draw-settings]]
+      [:div.center
+       [scramble-preview (:scramble-algorithm @state)]]
+      [settings-component state]])))