+++ src/squanmate/services/color_chooser.cljs
@@ -21,20 +21,3 @@
          :back (:front cs)
          :left (:right cs)
          :right (:left cs)))
-
-(defn make-color-settings
-  "Takes a color chooser state as a map and returns a ColorSettings instance
-  that can be used to get to the actual colors"
-  [{:keys [monochrome?
-           use-back-as-front
-           swap-top-and-bottom]
-    :or {monochrome? true}}]
-  (cond-> color-settings/defaults
-    monochrome?
-    make-monochrome
-
-    use-back-as-front
-    turn-y2
-
-    swap-top-and-bottom
-    do-swap-top-and-bottom))