+++ test/squanmate/alg/execution_test.cljs
@@ -57,8 +57,7 @@
    "this should do an alg and then reverse it. so it should show a solved
    puzzle. It uses the same alg as the test below, which is: "
    sample-alg
-   [newmonochrome/monochrome-puzzle (e/puzzle-of-result (apply-alg-and-its-reverse-alg sample-alg))
-    {:monochrome? false}]])
+   [newmonochrome/monochrome-puzzle (e/puzzle-of-result (apply-alg-and-its-reverse-alg sample-alg))]])
 
 (deftest transformation-result-reverse-test []
   (is (= puzzle/square-square